Fix: wait for complete model before getting props
When include in other project, the filer module content model is loaded
before the model of the project. Thus, on initialization, the filer
module could not fetch project aspects inheriting of filer
propertyInheritance aspect. This result in disabling property
inheritance.

Classes that need to fetch information from the complete content model
register themselves as DictionaryListener and process after the
dictionary initalization.

Change-Id: Idf2ef00bcf70fe57bafd71433a3d1a75d9ec1e2c