Interface for application context aware tests
Test classes that need to load the application context can implement the
interface.

It load all the application context and test context files.

All new test context files must be added to the interface's context
configuration. This is to ensure that each test share the same context,
thus loading the context once for all tests.

Change-Id: Ie266930250b516782ae73908e7e3aea7916197b4