feat: allow not changing owner
Every fileable's owner was set to filer.owner.username value. Now it is
possible to empty the filer.owner.username property to avoid changing
the owner.

Default value is still set to admin in the Alfresco Filer module so one
need to override the filer.owner.username property in the
alfresco-global.properties file in order to alter the default behaviour.

Change-Id: I80cb1bed31ff4044f1039be7fc469b72ba7d9a94